requirements.txt
streamlit
pandas
openai
python-dotenv
scikit-learn
nltk
